add a script for monipulating iOS device tree for running in qemu
This script reads the original device tree from iOS software update
and adds/changes the relevant nodes/properties to run in qemu.

First it removes most devices from being identified by iokit to load
new drivers. We don't want the drivers to load to not crash the kernel
as the devices themselves are not implemented yet in qemu.

We remove the "secure-root-prefix" prop as the kernel code hangs
before executing usermode apps if it is present.

We remove the upper bit in prop size which is the flag:
DT_PROP_FLAG_PLACEHOLDER. This removal process is usually done
by iBoot which we do not run.

We add some other prop values which are required for the kernel
and some kexts to run.